refactor(hooks): clean up the five #617 inline-disable sites
Each of the five remaining TODO(#617) markers resolved per the chosen
strategy. Five sites, five separate fixes bundled because they're
small and the theme is unified.

1. DangerZone.tsx:573  | rule: react-hooks/set-state-in-effect
   Initial async data loads on mount are the standard React pattern;
   the rule is overzealous here. Disable kept; TODO marker dropped
   and replaced with an explanation of why the rule is wrong.

2. MainPage.tsx:195    | rule: react-hooks/set-state-in-effect
   Recovery of in-flight sync state on QAM re-mount is legitimate
   mount-time setState. Same treatment as DangerZone.

3. RomMPlaySection.tsx:579  | rule: react-hooks/immutability
   The module-scope _cachedGameDetailCache moved into a new
   utils/cachedGameDetailStore.ts module. Components now call
   invalidateCachedGameDetail(appId) instead of mutating the cache
   directly. backend.ts re-exports the new API for back-compat.
   Tests moved to cachedGameDetailStore.test.ts.

4. SavesTab.tsx:234    | typing: DialogButton as any
   Module-augmented @decky/ui's DialogButtonProps to include
   onFocus?: (e: FocusEvent) => void in
   src/types/decky-ui-augmentation.d.ts. Dropped the 'as any' cast.
   Also removed three now-unused '@ts-expect-error onFocus' comments
   in CustomPlayButton.tsx (x2) and MainPage.tsx.
   The react-hooks/refs disable remains — the rule fires
   independently on createElement-of-forwardRef in conditional
   positions; the comment now explains why.

5. saves/SlotPanel.tsx:315  | rule: react-hooks/refs
   Promoted renderInactiveSlotBody to a proper <InactiveSlotBody>
   component in src/components/saves/InactiveSlotBody.tsx. Owns its
   own romm-slot-body wrapper div. SlotPanel.tsx body assignment
   simplified to a single ReactElement (no more spread arrays).
   A separate react-hooks/refs disable was added to the SlotPanel
   ternary that selects between active/inactive bodies — the rule
   fires on createElement-of-FC in ternary branches independent of
   the component's actual ref shape.

Net warning count: 60 → 56 (the three @ts-expect-error comments and
the DialogButton 'as any' cast accounted for the four reductions).
