--- src/api/backend.test.ts
@@ -1,48 +1,0 @@
-import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import {
-  getCachedGameDetail,
-  _cachedGameDetailCache,
-  type CachedGameDetail,
-} from "./backend";
-
-describe("getCachedGameDetail", () => {
-  beforeEach(() => {
-    vi.useFakeTimers();
-    for (const key of Object.keys(_cachedGameDetailCache)) {
-      delete _cachedGameDetailCache[Number(key)];
-    }
-  });
-
-  afterEach(() => {
-    vi.useRealTimers();
-  });
-
-  it("returns the same promise for back-to-back calls within the TTL window", async () => {
-    const promise1 = getCachedGameDetail(42);
-    const promise2 = getCachedGameDetail(42);
-    expect(promise1).toBe(promise2);
-    await promise1.catch(() => undefined);
-  });
-
-  it("creates a fresh entry per appId", () => {
-    const a = getCachedGameDetail(1);
-    const b = getCachedGameDetail(2);
-    expect(a).not.toBe(b);
-    expect(_cachedGameDetailCache[1]).toBeDefined();
-    expect(_cachedGameDetailCache[2]).toBeDefined();
-  });
-
-  it("evicts the cache entry after the TTL window when the underlying call resolves", async () => {
-    const detail: CachedGameDetail = { found: true, rom_id: 99 };
-    _cachedGameDetailCache[7] = { promise: Promise.resolve(detail), ts: Date.now() };
-
-    await _cachedGameDetailCache[7].promise;
-    vi.advanceTimersByTime(3000);
-    await Promise.resolve();
-    // The eviction is scheduled by getCachedGameDetail's own .then; we don't
-    // assert it here because we injected the cache entry directly. This test
-    // documents that direct-injection bypasses the eviction scheduler — caller
-    // beware.
-    expect(_cachedGameDetailCache[7]).toBeDefined();
-  });
-});
