+++ src/components/RomMPlaySection.tsx
@@ -30,7 +30,7 @@ import { scrollToTop } from "../utils/scrollHelpers";
 import { getEventTarget } from "../utils/events";
 import {
   getCachedGameDetail,
-  _cachedGameDetailCache,
+  invalidateCachedGameDetail,
   testConnection,
   getSaveStatus,
   getBiosStatus,
@@ -576,8 +576,7 @@ export const RomMPlaySection: FC<RomMPlaySectionProps> = ({ appId }) => { // NOS
           }));
         }
         // Invalidate the frontend cache and notify other components (e.g. GameInfoPanel)
-        // eslint-disable-next-line react-hooks/immutability -- TODO(#617): replace mutable cache with a setState-backed cache
-        delete (_cachedGameDetailCache as Record<number, unknown>)[appId];
+        invalidateCachedGameDetail(appId);
         globalThis.dispatchEvent(new CustomEvent("romm_data_changed", { detail: { type: "core_changed", platform_slug: info.platformSlug } }));
       } else {
         toaster.toast({ title: "RomM Sync", body: result.message || "Failed to set core" });
